+++ CHANGELOG.md
@@ -5,13 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
 
-## [1.27.2] - 2023-10-06
+## [Unreleased]
 
 ### Added
 
-- Show the upload modal when the user selects bulk import
+- Add Bulk import feature
 
 ## [1.28.2] - 2023-12-19